0497.02 subir bajar temperatura

+++ aireAcondicionado.java
@@ -1,17 +1,17 @@
 public class aireAcondicionado{
     // Temperatura del agua
-    private float temperatura;
+    private double temperatura;
     
     public aireAcondicionado(){
-        temperatura = 15.0f;
+        temperatura = 15.0;
     }
     public void subirTemperatura(){
-        temperatura += 5.0f;
+        temperatura += 5.0;
     }
     public void bajarTemperatura(){
-        temperatura -= 5.0f;
+        temperatura -= 5.0;
     }
-    public float getTemperatura(){
+    public double getTemperatura(){
         return temperatura;       
     }
 }
